Breakout TestWriters.swift into individual files.
