+++ Tests/TraceLogTests/TestWriters/SleepyTestWriter.swift
@@ -1,0 +1,38 @@
+///
+///  SleepyTestWriter.swift
+///
+///  Copyright 2018 Tony Stone
+///
+///  Licensed under the Apache License, Version 2.0 (the "License");
+///  you may not use this file except in compliance with the License.
+///  You may obtain a copy of the License at
+///
+///  http://www.apache.org/licenses/LICENSE-2.0
+///
+///  Unless required by applicable law or agreed to in writing, software
+///  distributed under the License is distributed on an "AS IS" BASIS,
+///  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+///  See the License for the specific language governing permissions and
+///  limitations under the License.
+///
+///  Created by Tony Stone on 6/20/18.
+///
+import XCTest
+import Foundation
+
+import TraceLog
+
+///
+/// Null Test Writer that sleeps for the passed in time when a message is recieved.
+///
+class SleepyTestWriter: Writer {
+    let sleepTime: useconds_t
+
+    init(sleepTime: useconds_t) {
+        self.sleepTime = sleepTime
+    }
+
+    func log(_ timestamp: Double, level: LogLevel, tag: String, message: String, runtimeContext: RuntimeContext, staticContext: StaticContext) {
+        usleep(sleepTime)
+    }
+}
